+++ CHANGELOG.md
@@ -9,7 +9,8 @@ This project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.htm
 
 ### Added
 
-- Add Npgsql and MySql.Data traces instrumentation.
+- Add MySql.Data traces instrumentation.
+- Add Npgsql traces instrumentation.
 - Add configuration option `none` to `OTEL_DOTNET_AUTO_TRACES_ENABLED_INSTRUMENTATIONS`
   and `OTEL_DOTNET_AUTO_METRICS_ENABLED_INSTRUMENTATIONS`.